Assign Adult Major Conditions cohort

A person is considered to be in this cohort if their age is over 18 and the recid is 01B, or their prescribing cost is £500 or over

Usage

assign_adult_major_cohort(recid, age, cost_total_net)

Arguments

recid

A vector of record IDs

age

A vector of ages

cost_total_net

A vector of total net costs

Value

A boolean vector indicating whether a given record is in the Adult Major Conditions cohort

See also