Assign a flag for elective instances

An elective instance is defined as when the CIJ patient type is Elective and the IPDC marker is D

Usage

assign_elective_instances(cij_pattype, cij_ipdc)

Arguments

cij_pattype

A vector of CIJ patient types

cij_ipdc

A vector of CIJ IPDC markers

Value

A boolean vector showing whether a record is an elective case or not

See also