Assign Frailty cohort

A record is considered to be in the frailty cohort if:

The recid is 01B, 50B, 02B, 04B or AE2 and

One of the diagnosis codes starts with W0 or W1

One of the diagnosis codes starts with F00, F01, F02, F03, F05, I61, I63, I64, G20 or G21

One of the diagnosis codes starts with R268 or G22X

The specialty is AB

The significant facility is 1E or 1D

The recid is GLS

Usage

assign_frailty_cohort(
  recid,
  diag1,
  diag2,
  diag3,
  diag4,
  diag5,
  diag6,
  spec,
  sigfac
)

Arguments

recid

A vector of record IDs

diag1, diag2, diag3, diag4, diag5, diag6

Character vectors of ICD-10 diagnosis codes.

spec

A vector of specialty codes

sigfac

A vector of significant facilities

Value

A boolean vector indicating whether a given record is in the Frailty cohort

See also