Generate counts of beddays per month for an episode with an admission and discharge date

Generate counts of beddays per month for an episode with an admission and discharge date

create_monthly_beddays(
  data,
  year,
  admission_date,
  discharge_date,
  include_costs = FALSE,
  count_last = TRUE
)

Arguments

data

Data to calculate beddays for.

year

The financial year in '1718' format.

admission_date

The admission/start date variable.

discharge_date

The admission/start date variable

include_costs

(default FALSE) - Select true for calculating monthly costs

count_last

(default TRUE) - Should the last day be counted, instead of the first?

Value

The data with additional variables apr_beddays to mar_beddaysthat count the beddays which occurred in the month.