Return the next financial year

Return the next financial year

Usage

next_fy(year, format = c("fyyear", "alternate"))

Arguments

year

a character vector of years

format

the format of the year vector, default is financial year

Value

A character of the date in the requested format

See also

Examples

next_fy("1718")
#> [1] "1819"