Previous update

Get the date of the previous update, e.g 'Mar_2022'

Usage

previous_update(months_ago = 3L, override = NULL)

Arguments

months_ago

Number of months since the previous update the default is 3 i.e. one quarter ago.

override

This allows specifying a specific update month if required.

Value

previous update as MMM_YYYY

See also

Examples

previous_update() # Default 3 months
#> Dec_2024
previous_update(1) # 1 month ago
#> Feb_2025
previous_update(override = "May_2023") # Specific Month
#> [1] "May_2023"