Process the Acute extract

This will read and process the acute extract, it will return the final data and (optionally) write it to disk.

Usage

process_extract_acute(
  data,
  year,
  acute_cup_path = get_boxi_extract_path(year, "acute_cup"),
  write_to_disk = TRUE
)

Arguments

data

The extract to process

year

The year to process, in FY format.

acute_cup_path

path to acute_cup data

write_to_disk

(optional) Should the data be written to disk default is TRUE i.e. write the data to disk.

Value

the final data as a tibble.

See also