Process the (year specific) Alarms Telecare extract

This will read and process the (year specific) Alarms Telecare extract, it will return the final data but also write this out as rds.

Usage

process_extract_alarms_telecare(
  data = NULL,
  year,
  client_lookup = NULL,
  write_to_disk = TRUE
)

Arguments

data

The extract to process. (Optional) Can be passed through a data list or alternatively read the file from disk.

year

The year to process, in FY format.

client_lookup

The client lookup extract (Optional) Can be passed through a data list or alternatively read the file from disk.

write_to_disk

(optional) Should the data be written to disk default is TRUE i.e. write the data to disk.

Value

the final data as a tibble.

See also