Source Care Home Tests

Produce a set of tests which can be used by Care Home and Home Care This will produce counts of various demographics using create_demog_test_flags() counts of episodes for every hbtreatcode using create_lca_test_flags() It will also produce various summary statistics for bedday, cost and episode date variables.

Usage

produce_source_ch_tests(
  data,
  sum_mean_vars = c("beddays", "cost", "yearstay"),
  max_min_vars = c("record_keydate1", "record_keydate2", "cost_total_net", "yearstay")
)

Arguments

data

new or old data for testing summary flags (data is from get_source_extract_path())

sum_mean_vars

variables used when selecting 'all' measures from calculate_measures()

max_min_vars

variables used when selecting 'min-max' from calculate_measures()

Value

a dataframe with a count of each flag from calculate_measures()

See also