Write a data to a file

Write data to a file, the function chosen to write the file is dependant on the file path extension.

.rds uses readr::write_rds().

.parquet uses arrow::write_parquet().

Usage

write_file(data, path, ...)

Arguments

data

The data to be written

path

The file path to be write

...

Additional arguments passed to the relevant function.

Value

the data (invisibly) as a tibble.