Get and check and full file path

This generic function takes a directory and file name then checks to make sure they exist. The parameter check_mode will also test to make sure the file is readable (default) or writeable (check_mode = "write"). By default it will return an error if the file doesn't exist but with create = TRUE it will create an empty file with appropriate permissions.

get_file_path(
  directory,
  file_name,
  ext = NULL,
  check_mode = "read",
  create = NULL
)

Arguments

directory

The file directory

file_name

The file name (with extension if not supplied to ext)

ext

The extension (type of the file) - optional

check_mode

The mode passed to file_access, defaults to "read" to check that you have read access to the file

create

Optionally create the file if it doesn't exists, the default is to only create a file if we set check_mode = "write"

Value

The full file path, an error will be thrown if the path doesn't exist or it's not readable

See also